test-xooj/requirement.txt
django==1.10.5
redis==2.10.6
python-memcached==1.58
django-filter==1.0.1
djangorestframework==3.5.4
mysqlclient==1.3.10
pyquery==1.2.17
psutil==5.2.2
enum==0.4.6
django-cached-authentication-middleware==0.2.2
python-openstackclient==3.14.1
python-novaclient==9.1.1
python-neutronclient==6.7.0
python-glanceclient==2.10.0
python-keystoneclient==3.15.0
python-cinderclient==3.5.0
python-ceilometerclient==2.9.0
python-zunclient==1.1.0
docker==3.4.1
python-docx==0.8.6
Pillow==4.1.1
cryptography==2.0.3
xlrd==1.1.0
xlsxwriter==1.0.2
reportlab==3.4.0
PyPDF2==1.26.0
daphne==1.1.0
channels==1.0.3
zio==1.0.3
requests==2.13.0
pwntools==3.9.2
asgiref==1.1.2
asgi-redis==1.0.0
paramiko==2.1.2
py-gfm==0.1.3
markdown==2.6.11
retry==0.9.2
mkdocs==0.17.2
ffmpy==0.2.2
service_identity==17.0.0
pyminizip==0.2.1
thriftpy==0.3.9
moviepy==0.2.3.2
pymdown-extensions==4.10.1
mkdocs-material==2.7.3
xlwt==1.3.0
pymysql==0.9.3
ConcurrentLogHandler==0.9.1
Cython==0.29.5
django-redis==4.4.1
